refactor first-post-published-modal
